Added Oauth 2.0 functionality for the app
