+++ app/api/auth/[...nextauth]/route.js
@@ -1,0 +1,22 @@
+import NextAuth from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+
+const handler = NextAuth({
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+    }),
+  ],
+  session: {
+    strategy: "jwt",
+  },
+  callbacks: {
+    async session({ session, token }) {
+      session.userId = token.sub; // Attach user's OAuth ID to the session
+      return session;
+    },
+  },
+});
+
+export { handler as GET, handler as POST };
