+++ app/layout.tsx
@@ -1,17 +1,16 @@
+"use client";
+
 import type { Metadata } from "next";
 import { Geist, Geist_Mono, JetBrains_Mono } from "next/font/google";
 import "./globals.css";
 import { Poppins } from "next/font/google";
+import { SessionProvider } from "next-auth/react";
 
 const poppins = Poppins({
   subsets: ["latin"],
   weight: ["400", "600"], // customize weights as needed
 });
 
-export const metadata: Metadata = {
-  title: "Latex Formatter",
-  description: "Generated with ❤️",
-};
 const geistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
@@ -34,7 +33,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={jetBrains_Mono.variable}>{children}</body>
+      <body className={jetBrains_Mono.variable}>
+        <SessionProvider>{children}</SessionProvider>
+      </body>
     </html>
   );
 }
