+++ app/metadata.ts
@@ -1,0 +1,6 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Latex Formatter",
+  description: "Generated with ❤️",
+};
