+++ package.json
@@ -18,6 +18,7 @@
     "latest": "^0.2.0",
     "n": "^10.1.0",
     "next": "15.3.0",
+    "next-auth": "^4.24.11",
     "node": "^23.11.0",
     "pg": "^8.15.5",
     "react": "^19.0.0",
